♻ ✅ ️ add attachmentService.ts, add tests for convertToExposedRecord
